requirements.txt
fastapi[standard]>=0.112.0,<0.114.0
pydantic>=2.7.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2>=2.9.9,<3.0.0
load_dotenv>=0.1.0,<0.2.0
PyJWT>=1.7.1,<3.0.0
sqladmin[full]>=0.18.0,<1.0.0